refactor: reorganize Section handler into dedicated subdirectory
- Move section_handler.go and section_handler_test.go to internal/handler/section/
- Update package declarations to 'section'
- Update imports in router and dependency injection
- Maintain all existing functionality and tests
- Improve code organization and maintainability
