+++ internal/application/di/section_dependencies.go
@@ -1,15 +1,15 @@
 package di
 
 import (
-	"ProyectoFinal/internal/handler"
+	sectionHandler "ProyectoFinal/internal/handler/section"
 	repository "ProyectoFinal/internal/repository/section"
 	service "ProyectoFinal/internal/service/section"
 	"database/sql"
 )
 
-func GetSectionHandler(db *sql.DB) *handler.SectionDefault {
+func GetSectionHandler(db *sql.DB) *sectionHandler.SectionDefault {
 	sectionRepository := repository.NewSectionMySQL(db)
 	sectionService := service.NewSectionDefault(sectionRepository)
-	sectionHandler := handler.NewSectionDefault(sectionService)
+	sectionHandler := sectionHandler.NewSectionDefault(sectionService)
 	return sectionHandler
 }
