+++ internal/application/router/section_router.go
@@ -2,11 +2,12 @@ package router
 
 import (
 	"ProyectoFinal/internal/handler"
+	sectionHandler "ProyectoFinal/internal/handler/section"
 
 	"github.com/go-chi/chi/v5"
 )
 
-func GetSectionRouter(handler *handler.SectionDefault, productBatchHandler *handler.ProductBatchHandler) chi.Router {
+func GetSectionRouter(handler *sectionHandler.SectionDefault, productBatchHandler *handler.ProductBatchHandler) chi.Router {
 	r := chi.NewRouter()
 	r.Post("/", handler.Create())
 	r.Get("/", handler.GetAll())
