+++ internal/handler/section/section_handler_test.go
@@ -1,4 +1,4 @@
-package handler
+package section
 
 import (
 	"ProyectoFinal/mocks"
